Extracting some of the generic test magic to ChimeTestCase. Making the magic more magic in that it should find all test methods.
